feat(src/Types): Add global types and import only enums from enums file
