+++ src/enums.ts
@@ -1,0 +1,24 @@
+export enum Web3Errors {
+  UNKNOWN_ERROR,
+  SIGNATURE_REJECTED,
+}
+export enum TXEvents {
+  TX_HASH = 'txhash',
+  RECEIPT = 'receipt',
+  CONFIRMATION = 'confirmation',
+  TX_ERROR = 'txerror',
+  FINALLY = 'finally',
+  INVARIANT = 'invariant',
+}
+
+export enum VoteDecision {
+  None,
+  YES,
+  NO
+}
+
+export enum WalletSchemeProposalState { Submitted, Passed, Failed, Executed }
+
+export enum VotingMachineProposalState { 
+  None, ExpiredInQueue, Executed, Queued, PreBoosted, Boosted, QuietEndingPeriod
+}
